requirements.txt
fastapi
uvicorn
python-multipart
ollama
opencv-python
numpy
tensorflow
Pillow
requests
tqdm